Buscar funcionário ok

+++ PetShop/Model/Funcionario.cs
@@ -9,9 +9,9 @@ namespace PetShop.Model
     class Funcionario : Pessoa
     {
         private string carttrab;
-        private double salario;
+        private decimal salario;
 
         public string Carttrab { get => carttrab; set => carttrab = value; }
-        public double Salario { get => salario; set => salario = value; }
+        public decimal Salario { get => salario; set => salario = value; }
     }
 }
